+++ CHANGELOG.md
@@ -831,6 +831,8 @@ All notable changes to this project will be documented in this file.
 - Add arcade meta-system
 - Add amiga emulator (libretro-uae) on all platforms
 
+- Add a script to install currently running Recalbox on another storage device
+
 ## [6.0] - DragonBlaze
 - Change name from 5.0 to 6.0 DragonBlaze
 - Adding GPIO, Custom GPIO and MCP inputConfig in es_input.cfg